src/SurveillanceAgent/requirements.txt
opencv-python==4.8.1.78
numpy>=1.21.0
Pillow>=9.0.0
requests>=2.25.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0
dataclasses-json>=0.5.0
pydantic>=2.0.0
loguru>=0.6.0
tenacity>=8.0.0
aiohttp>=3.8.0
pydantic-settings>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
